begin refactoring LSP portion
