+++ src/goMain.ts
@@ -6,18 +6,7 @@
 'use strict';
 
 import vscode = require('vscode');
-import { GoCompletionItemProvider, getCompletionsWithoutGoCode } from './goSuggest';
-import { GoHoverProvider } from './goExtraInfo';
-import { GoDefinitionProvider } from './goDeclaration';
-import { GoReferenceProvider } from './goReferences';
-import { GoImplementationProvider } from './goImplementations';
-import { GoTypeDefinitionProvider } from './goTypeDefinition';
-import { GoDocumentFormattingEditProvider } from './goFormat';
-import { GoRenameProvider } from './goRename';
-import { GoDocumentSymbolProvider } from './goOutline';
 import { GoRunTestCodeLensProvider } from './goRunTestCodelens';
-import { GoSignatureHelpProvider } from './goSignature';
-import { GoWorkspaceSymbolProvider } from './goSymbol';
 import { GoCodeActionProvider } from './goCodeAction';
 import { check, removeTestStatus, notifyIfGeneratedFile } from './goCheck';
 import { updateGoPathGoRootFromConfig, offerToInstallTools, promptForMissingTool, installTools } from './goInstallTools';
@@ -28,18 +17,12 @@ import { testAtCursor, testCurrentPackage, testCurrentFile, testPrevious, testWo
 import { showTestOutput, cancelRunningTests } from './testUtils';
 import * as goGenerateTests from './goGenerateTests';
 import { addImport, addImportToWorkspace } from './goImport';
-import { installAllTools, getLanguageServerToolPath } from './goInstallTools';
+import { installAllTools } from './goInstallTools';
 import {
 	isGoPathSet, getBinPath, sendTelemetryEvent, getExtensionCommands, getGoVersion, getCurrentGoPath,
-	getToolsGopath, handleDiagnosticErrors, disposeTelemetryReporter, getToolsEnvVars, cleanupTempDir
+	getToolsGopath, handleDiagnosticErrors, disposeTelemetryReporter, cleanupTempDir
 } from './util';
-import {
-	LanguageClient, RevealOutputChannelOn, FormattingOptions, ProvideDocumentFormattingEditsSignature,
-	ProvideCompletionItemsSignature, ProvideRenameEditsSignature, ProvideDefinitionSignature, ProvideHoverSignature,
-	ProvideReferencesSignature, ProvideSignatureHelpSignature, ProvideDocumentSymbolsSignature, ProvideWorkspaceSymbolsSignature,
-	HandleDiagnosticsSignature, ProvideDocumentLinksSignature,
-} from 'vscode-languageclient';
-import { clearCacheForTools, fixDriveCasingInWindows, getToolFromToolPath } from './goPath';
+import { clearCacheForTools, fixDriveCasingInWindows } from './goPath';
 import { addTags, removeTags } from './goModifytags';
 import { runFillStruct } from './goFillStruct';
 import { parseLiveFile } from './goLiveErrors';
@@ -55,8 +38,7 @@ import { vetCode } from './goVet';
 import { buildCode } from './goBuild';
 import { installCurrentPackage } from './goInstall';
 import { setGlobalState } from './stateUtils';
-import { ProvideTypeDefinitionSignature } from 'vscode-languageclient/lib/typeDefinition';
-import { ProvideImplementationSignature } from 'vscode-languageclient/lib/implementation';
+import { maybeRegisterLanguageServer } from './goLanguageServer';
 
 export let buildDiagnosticCollection: vscode.DiagnosticCollection;
 export let lintDiagnosticCollection: vscode.DiagnosticCollection;
@@ -106,218 +88,15 @@ export function activate(ctx: vscode.ExtensionContext): void {
 		ctx.globalState.update('toolsGoInfo', toolsGoInfo);
 
 		offerToInstallTools();
-		const languageServerToolPath = getLanguageServerToolPath();
-		if (languageServerToolPath) {
-			const languageServerTool = getToolFromToolPath(languageServerToolPath);
-			const languageServerExperimentalFeatures: any = vscode.workspace.getConfiguration('go').get('languageServerExperimentalFeatures') || {};
-			const langServerFlags: string[] = vscode.workspace.getConfiguration('go')['languageServerFlags'] || [];
 
-			// `-trace was a flag for `go-langserver` which is replaced with `-rpc.trace` for gopls to avoid crash
-			const traceFlagIndex = langServerFlags.indexOf('-trace');
-			if (traceFlagIndex > -1 && languageServerTool === 'gopls') {
-				langServerFlags[traceFlagIndex] = '-rpc.trace';
-			}
-
-			const c = new LanguageClient(
-				languageServerTool,
-				{
-					command: languageServerToolPath,
-					args: ['-mode=stdio', ...langServerFlags],
-					options: {
-						env: getToolsEnvVars()
-					}
-				},
-				{
-					initializationOptions: {
-						funcSnippetEnabled: vscode.workspace.getConfiguration('go')['useCodeSnippetsOnFunctionSuggest'],
-						gocodeCompletionEnabled: languageServerExperimentalFeatures['autoComplete'],
-						incrementalSync: languageServerExperimentalFeatures['incrementalSync']
-					},
-					documentSelector: ['go', 'go.mod', 'go.sum'],
-					uriConverters: {
-						// Apply file:/// scheme to all file paths.
-						code2Protocol: (uri: vscode.Uri): string => (uri.scheme ? uri : uri.with({ scheme: 'file' })).toString(),
-						protocol2Code: (uri: string) => vscode.Uri.parse(uri),
-					},
-					revealOutputChannelOn: RevealOutputChannelOn.Never,
-					middleware: {
-						provideDocumentFormattingEdits: (document: vscode.TextDocument, options: FormattingOptions, token: vscode.CancellationToken, next: ProvideDocumentFormattingEditsSignature) => {
-							if (languageServerExperimentalFeatures['format'] === true) {
-								return next(document, options, token);
-							}
-							return [];
-						},
-						provideCompletionItem: async (document: vscode.TextDocument, position: vscode.Position, context: vscode.CompletionContext, token: vscode.CancellationToken, next: ProvideCompletionItemsSignature) => {
-							if (languageServerExperimentalFeatures['autoComplete'] === true) {
-								const promiseFromLanguageServer = Promise.resolve(next(document, position, context, token));
-								const promiseWithoutGoCode = getCompletionsWithoutGoCode(document, position);
-								const [resultFromLanguageServer, resultWithoutGoCode] = await Promise.all([promiseFromLanguageServer, promiseWithoutGoCode]);
-								if (!resultWithoutGoCode || !resultWithoutGoCode.length) {
-									return resultFromLanguageServer;
-								}
-								const completionItemsFromLanguageServer = Array.isArray(resultFromLanguageServer) ? resultFromLanguageServer : resultFromLanguageServer.items;
-								resultWithoutGoCode.forEach(x => {
-									if (x.kind !== vscode.CompletionItemKind.Module || !completionItemsFromLanguageServer.some(y => y.label === x.label)) {
-										completionItemsFromLanguageServer.push(x);
-									}
-								});
-								return resultFromLanguageServer;
-							}
-							return [];
-						},
-						provideRenameEdits: (document: vscode.TextDocument, position: vscode.Position, newName: string, token: vscode.CancellationToken, next: ProvideRenameEditsSignature) => {
-							if (languageServerExperimentalFeatures['rename'] === true) {
-								return next(document, position, newName, token);
-							}
-							return null;
-						},
-						provideDefinition: (document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, next: ProvideDefinitionSignature) => {
-							if (languageServerExperimentalFeatures['goToDefinition'] === true) {
-								return next(document, position, token);
-							}
-							return null;
-						},
-						provideTypeDefinition: (document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, next: ProvideTypeDefinitionSignature) => {
-							if (languageServerExperimentalFeatures['goToTypeDefinition'] === true) {
-								return next(document, position, token);
-							}
-							return null;
-						},
-						provideHover: (document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, next: ProvideHoverSignature) => {
-							if (languageServerExperimentalFeatures['hover'] === true) {
-								return next(document, position, token);
-							}
-							return null;
-						},
-						provideReferences: (document: vscode.TextDocument, position: vscode.Position, options: { includeDeclaration: boolean }, token: vscode.CancellationToken, next: ProvideReferencesSignature) => {
-							if (languageServerExperimentalFeatures['findReferences'] === true) {
-								return next(document, position, options, token);
-							}
-							return [];
-						},
-						provideSignatureHelp: (document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, next: ProvideSignatureHelpSignature) => {
-							if (languageServerExperimentalFeatures['signatureHelp'] === true) {
-								return next(document, position, token);
-							}
-							return null;
-						},
-						provideDocumentSymbols: (document: vscode.TextDocument, token: vscode.CancellationToken, next: ProvideDocumentSymbolsSignature) => {
-							if (languageServerExperimentalFeatures['documentSymbols'] === true) {
-								return next(document, token);
-							}
-							return [];
-						},
-						provideWorkspaceSymbols: (query: string, token: vscode.CancellationToken, next: ProvideWorkspaceSymbolsSignature) => {
-							if (languageServerExperimentalFeatures['workspaceSymbols'] === true) {
-								return next(query, token);
-							}
-							return [];
-						},
-						provideImplementation: (document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, next: ProvideImplementationSignature) => {
-							if (languageServerExperimentalFeatures['goToImplementation'] === true) {
-								return next(document, position, token);
-							}
-							return null;
-						},
-						handleDiagnostics: (uri: vscode.Uri, diagnostics: vscode.Diagnostic[], next: HandleDiagnosticsSignature) => {
-							if (languageServerExperimentalFeatures['diagnostics'] === true) {
-								return next(uri, diagnostics);
-							}
-							return null;
-						},
-						provideDocumentLinks: (document: vscode.TextDocument, token: vscode.CancellationToken, next: ProvideDocumentLinksSignature) => {
-							if (languageServerExperimentalFeatures['documentLink'] === true) {
-								return next(document, token);
-							}
-							return null;
-						}
-					}
-				}
-			);
-
-			c.onReady().then(() => {
-				const capabilities = c.initializeResult && c.initializeResult.capabilities;
-				if (!capabilities) {
-					return vscode.window.showErrorMessage('The language server is not able to serve any features at the moment.');
-				}
-
-				if (languageServerExperimentalFeatures['autoComplete'] !== true || !capabilities.completionProvider) {
-					registerCompletionProvider(ctx);
-				}
-
-				if (languageServerExperimentalFeatures['format'] !== true || !capabilities.documentFormattingProvider) {
-					ctx.subscriptions.push(vscode.languages.registerDocumentFormattingEditProvider(GO_MODE, new GoDocumentFormattingEditProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['rename'] !== true || !capabilities.renameProvider) {
-					ctx.subscriptions.push(vscode.languages.registerRenameProvider(GO_MODE, new GoRenameProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['goToTypeDefinition'] !== true || !capabilities.typeDefinitionProvider) {
-					ctx.subscriptions.push(vscode.languages.registerTypeDefinitionProvider(GO_MODE, new GoTypeDefinitionProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['hover'] !== true || !capabilities.hoverProvider) {
-					ctx.subscriptions.push(vscode.languages.registerHoverProvider(GO_MODE, new GoHoverProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['goToDefinition'] !== true || !capabilities.definitionProvider) {
-					ctx.subscriptions.push(vscode.languages.registerDefinitionProvider(GO_MODE, new GoDefinitionProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['findReferences'] !== true || !capabilities.referencesProvider) {
-					ctx.subscriptions.push(vscode.languages.registerReferenceProvider(GO_MODE, new GoReferenceProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['documentSymbols'] !== true || !capabilities.documentSymbolProvider) {
-					ctx.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(GO_MODE, new GoDocumentSymbolProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['signatureHelp'] !== true || !capabilities.signatureHelpProvider) {
-					ctx.subscriptions.push(vscode.languages.registerSignatureHelpProvider(GO_MODE, new GoSignatureHelpProvider(), '(', ','));
-				}
-
-				if (languageServerExperimentalFeatures['workspaceSymbols'] !== true || !capabilities.workspaceSymbolProvider) {
-					ctx.subscriptions.push(vscode.languages.registerWorkspaceSymbolProvider(new GoWorkspaceSymbolProvider()));
-				}
-
-				if (languageServerExperimentalFeatures['goToImplementation'] !== true || !capabilities.implementationProvider) {
-					ctx.subscriptions.push(vscode.languages.registerImplementationProvider(GO_MODE, new GoImplementationProvider()));
-				}
-
-			});
-
-			let languageServerDisposable = c.start();
-			ctx.subscriptions.push(languageServerDisposable);
-
-			ctx.subscriptions.push(vscode.commands.registerCommand('go.languageserver.restart', async () => {
-				await c.stop();
-				languageServerDisposable.dispose();
-				languageServerDisposable = c.start();
-				ctx.subscriptions.push(languageServerDisposable);
-			}));
-
-			if (languageServerTool !== 'gopls' || !languageServerExperimentalFeatures['diagnostics']) {
-				vscode.workspace.onDidChangeTextDocument(parseLiveFile, null, ctx.subscriptions);
-			}
-		} else {
-			registerCompletionProvider(ctx);
-			ctx.subscriptions.push(vscode.languages.registerHoverProvider(GO_MODE, new GoHoverProvider()));
-			ctx.subscriptions.push(vscode.languages.registerDefinitionProvider(GO_MODE, new GoDefinitionProvider()));
-			ctx.subscriptions.push(vscode.languages.registerReferenceProvider(GO_MODE, new GoReferenceProvider()));
-			ctx.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(GO_MODE, new GoDocumentSymbolProvider()));
-			ctx.subscriptions.push(vscode.languages.registerWorkspaceSymbolProvider(new GoWorkspaceSymbolProvider()));
-			ctx.subscriptions.push(vscode.languages.registerSignatureHelpProvider(GO_MODE, new GoSignatureHelpProvider(), '(', ','));
-			ctx.subscriptions.push(vscode.languages.registerImplementationProvider(GO_MODE, new GoImplementationProvider()));
-			ctx.subscriptions.push(vscode.languages.registerDocumentFormattingEditProvider(GO_MODE, new GoDocumentFormattingEditProvider()));
-			ctx.subscriptions.push(vscode.languages.registerTypeDefinitionProvider(GO_MODE, new GoTypeDefinitionProvider()));
-			ctx.subscriptions.push(vscode.languages.registerRenameProvider(GO_MODE, new GoRenameProvider()));
-			vscode.workspace.onDidChangeTextDocument(parseLiveFile, null, ctx.subscriptions);
-		}
+		// This handles all of the configurations and registrations for the language server.
+		// It also registers the necessary language feature providers that the language server may not support.
+		maybeRegisterLanguageServer(ctx);
 
 		if (vscode.window.activeTextEditor && vscode.window.activeTextEditor.document.languageId === 'go' && isGoPathSet()) {
 			runBuilds(vscode.window.activeTextEditor.document, vscode.workspace.getConfiguration('go', vscode.window.activeTextEditor.document.uri));
 		}
+
 	});
 
 	initCoverageDecorators(ctx);
@@ -483,36 +262,12 @@ export function activate(ctx: vscode.ExtensionContext): void {
 			clearCacheForTools();
 		}
 
-		const languageServerToolPathBeingUsed = getLanguageServerToolPath();
-		let reloadMessage: string;
-		if (e.affectsConfiguration('go.useLanguageServer')) {
-			if (updatedGoConfig['useLanguageServer']) {
-				if (languageServerToolPathBeingUsed) {
-					reloadMessage = 'Reload VS Code window to enable the use of language server';
-				}
-			} else {
-				reloadMessage = 'Reload VS Code window to disable the use of language server';
-			}
-		} else if (e.affectsConfiguration('go.languageServerFlags') || e.affectsConfiguration('go.languageServerExperimentalFeatures')) {
-			reloadMessage = 'Reload VS Code window for the changes in language server settings to take effect';
-		}
-
-		// If there was a change in "useLanguageServer" setting, then ask the user to reload VS Code.
-		if (reloadMessage) {
-			vscode.window.showInformationMessage(reloadMessage, 'Reload').then(selected => {
-				if (selected === 'Reload') {
-					vscode.commands.executeCommand('workbench.action.reloadWindow');
-				}
-			});
-		}
-
 		if (updatedGoConfig['enableCodeLens']) {
 			testCodeLensProvider.setEnabled(updatedGoConfig['enableCodeLens']['runtest']);
 			referencesCodeLensProvider.setEnabled(updatedGoConfig['enableCodeLens']['references']);
 		}
 
-		const languageServerExperimentalFeatures: any = updatedGoConfig.get('languageServerExperimentalFeatures');
-		if (e.affectsConfiguration('go.formatTool') && (!languageServerToolPathBeingUsed || languageServerExperimentalFeatures['format'] === false)) {
+		if (e.affectsConfiguration('go.formatTool')) {
 			checkToolExists(updatedGoConfig['formatTool']);
 		}
 		if (e.affectsConfiguration('go.lintTool')) {
@@ -735,10 +490,4 @@ function checkToolExists(tool: string) {
 	if (tool === getBinPath(tool)) {
 		promptForMissingTool(tool);
 	}
-}
-
-function registerCompletionProvider(ctx: vscode.ExtensionContext) {
-	const provider = new GoCompletionItemProvider(ctx.globalState);
-	ctx.subscriptions.push(provider);
-	ctx.subscriptions.push(vscode.languages.registerCompletionItemProvider(GO_MODE, provider, '.', '\"'));
 }
